+++ src/components/level/toolbars/EditToolbar/selection-editors/flip.tsx
@@ -1,5 +1,6 @@
 import { FlipDirection, getDriver } from "drivers";
 import { $currentDriverName } from "models/levelsets";
+import { HotKeyMask } from "models/ui/hotkeys";
 import { svgs } from "ui/icon";
 import { IBounds, Point2D } from "utils/rect";
 import { SelectionEditor, SelectionEditorFn } from "./_types";
@@ -36,6 +37,7 @@ export const flipH: SelectionEditor = {
     (v, c) => ({ x: v, y: c }),
     FlipDirection.H,
   ),
+  hotkeys: ["H", HotKeyMask.SHIFT],
 };
 export const flipV: SelectionEditor = {
   title: "Flip Vertical",
@@ -45,4 +47,5 @@ export const flipV: SelectionEditor = {
     (v, c) => ({ x: c, y: v }),
     FlipDirection.V,
   ),
+  hotkeys: ["V", HotKeyMask.SHIFT],
 };