+++ src/components/level/toolbars/EditToolbar/selection-editors/rnd.tsx
@@ -5,6 +5,7 @@ import { FC, useCallback, useMemo } from "react";
 import * as RoMap from "@cubux/readonly-map";
 import { getDriver, getTilesForToolbar } from "drivers";
 import { $currentDriverName } from "models/levelsets";
+import { HotKeyMask } from "models/ui/hotkeys";
 import { Button, TextButton } from "ui/button";
 import { svgs } from "ui/icon";
 import { Field, Range } from "ui/input";
@@ -154,4 +155,5 @@ export const rnd: SelectionEditor = {
   title: "Random",
   icon: <svgs.Random />,
   Component: RndEditor,
+  hotkeys: ["N", HotKeyMask.SHIFT],
 };