+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.11.0 (DEV)
 
 - Add: "Straight line" drawing tool.
+- Add: Keyboard shortcuts to Selection Editors from previous release.
 - Enh: Replace number inputs and selects in Maze and Random tools with sliders.
 - Enh: Selection Editors now remember theirs values.